// src/main/webapp/app/entities/client/client.model.ts
import { IClientPhoto } from 'app/entities/client-photo/client-photo.model';
import { IClientVenueStats } from 'app/entities/client-venue-stats/client-venue-stats.model';

export interface IClient {
  id: string;
  clientId?: string | null;
  createdDate?: string | null;
  updatedDate?: string | null;
  deletedDate?: string | null;
  lastname?: string | null;
  firstname?: string | null;
  gender?: string | null;
  salutation?: string | null;
  title?: string | null;
  birthdayDay?: number | null;
  birthdayMonth?: number | null;
  birthdayAltMonth?: number | null;
  anniversaryDay?: number | null;
  anniversaryMonth?: number | null;
  company?: string | null;
  email?: string | null;
  emailAlt?: string | null;
  phoneNumber?: string | null;
  phoneNumberlocale?: string | null;
  phoneNumberalt?: string | null;
  phoneNumberaltlocale?: string | null;
  address?: string | null;
  address2?: string | null;
  city?: string | null;
  postalCode?: string | null;
  state?: string | null;
  country?: string | null;
  isContactPrivate?: boolean | null;
  isOnetimeGuest?: boolean | null;
  status?: string | null;
  loyaltyId?: string | null;
  loyaltyRank?: number | null;
  loyaltyTier?: string | null;
  marketingOptin?: boolean | null;
  marketingOptints?: string | null;
  marketingOptOutts?: string | null;
  hasBillingProfile?: boolean | null;
  notes?: string | null;
  privateNotes?: string | null;
  tags?: string | null;
  totalVisits?: number | null;
  totalCovers?: number | null;
  totalCancellations?: number | null;
  totalNoShows?: number | null;
  totalSpend?: number | null;
  totalSpendPerCover?: number | null;
  totalspendPerVisit?: number | null;
  avgRating?: number | null;
  referenceCode?: string | null;
  externalUserId?: string | null;
  venueGroupId?: string | null;
  birthdayAltDay?: number | null;
  userId?: string | null;
  userName?: string | null;
  totalOrderCount?: number | null;
  preferredLanguageCode?: string | null;
  clientPhoto?: IClientPhoto | null;
  clientVenueStats?: IClientVenueStats | null;
}

export type NewClient = Omit<IClient, 'id'> & { id: null };